% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{add_phylopic_to_map}
\alias{add_phylopic_to_map}
\title{Fixes error in adding phylopics to maps}
\usage{
add_phylopic_to_map(img, alpha = 0.2, x = NULL, y = NULL,
  ysize = NULL, color = NULL)
}
\description{
Get this error:
Error: annotation_custom only works with Cartesian coordinates
Using rphylopic::add_phylopic
This follows the advice of https://github.com/oswaldosantos/ggsn/issues/19
And just uses ggmap::inset instead; otherwise, it's identical to rphylopic::add_phylopic
}
