% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{add_points}
\alias{add_points}
\title{function to add pbdb paleo data points (lat and long) to gplatesr created maps}
\usage{
add_points(map, df)
}
\arguments{
\item{map}{a map created using gplatesr plot_gplates or black_white functions}

\item{df}{a data frame containing pbdb_data.paleolng and pbdb_data.paleolat}

\item{ptcolor}{a character vector indicating the desired color of added data points}
}
\value{
a map with points
}
\description{
function to add pbdb paleo data points (lat and long) to gplatesr created maps
}
