% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{AddAxis}
\alias{AddAxis}
\title{function to add axis to a tree plot}
\usage{
AddAxis(focalPeriod = NULL, age_df = GetAgeDF())
}
\arguments{
\item{focalPeriod}{The period to highlight, using names form GetAgeDF()}

\item{age_df}{Return from GetAgeDF(), which has colors}
}
\value{
modifies current plot
}
\description{
It can highlight chosen periods
}
