% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{amb}
\alias{amb}
\title{flips the direction of the tree at each node}
\usage{
amb(phy)
}
\arguments{
\item{phy}{an object of class phylo}
}
\value{
a phylo object
}
\description{
flips the direction of the tree at each node
}
