% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animatedmaps}
\alias{animatedmaps}
\title{Animated maps}
\format{A list of lists of outputs of AnimatePlots}
\usage{
data(animatedmaps)
}
\description{
List containing animated maps. It's a list of lists: overall list has "all", "none", and then individual taxa. Then, each of those has "all", and then individual periods. For example, you could get the maps for Stegosaurus in the Jurassic by doing animatedmaps[["Stegosaurus"]][["Jurassic"]]
}
\details{
Each element in that list of lists is itself a list, the output of AnimatePlot()
}
\keyword{datasets}
