% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{AnimatePlot}
\alias{AnimatePlot}
\title{Create an animated gif of a map}
\usage{
AnimatePlot(start_time = NULL, stop_time = NULL, periods = NULL,
  taxa = NULL, step_size = 1, age_df = GetAgeDF(),
  specimen_df = specimens, interval = 0.5,
  use_cached_maps_only = FALSE, use_phylopics = FALSE,
  point_color = "red", gif_name = NULL, paleomaps_allages = NULL,
  single_frame = FALSE)
}
\arguments{
\item{start_time}{The time of the first frame of the animation}

\item{stop_time}{The time of the last frame of the animation before it starts looping back}

\item{periods}{A vector of period names, capitalized properly (can be left blank)}

\item{taxa}{A vector of taxon names (can be left blank)}

\item{step_size}{How many million years to take in a single step}

\item{age_df}{Data.frame of ages, typically from GetAgeDF()}

\item{specimen_df}{Cached fossil localities and times}

\item{interval}{How many seconds per frame}

\item{use_cached_maps_only}{If TRUE, only uses the maps already in the package, rather than pulling from gplates}

\item{use_phylopics}{If TRUE, use phylopic images; otherwise, use dots}

\item{point_color}{If just plotting points, what color}

\item{gif_name}{Path to gif, including its name}

\item{paleomaps_allages}{Cache of maps}

\item{single_frame}{If TRUE, just does a single frame. It picks the best time.}
}
\description{
This can work with or without taxa.
}
\details{
The age range to plot can be set by the periods, the taxa, or fixed ages (which by default go from 0 to 540 MY). If taxa are specified, it uses the times those are found. If periods are specified, it uses the start and stop of those periods. If both periods and taxa are specified, it defaults to using the periods.

a <- ot(use_cached_maps_only=TRUE, step_size=1, taxa=GetTaxa())
}
