% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{cache_pictree}
\alias{cache_pictree}
\title{Cache phylotree tree information}
\usage{
cache_pictree(tree = GetTree())
}
\arguments{
\item{taxa}{vector of taxa}
}
\description{
Cache phylotree tree information
}
