% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{CacheAnimatedMaps}
\alias{CacheAnimatedMaps}
\title{Cache animated maps}
\usage{
CacheAnimatedMaps(start_time = NULL, stop_time = NULL,
  periods = NULL, taxa = GetTaxa(), step_size = 1,
  age_df = GetAgeDF(), specimen_df = specimens, interval = 0.5,
  use_cached_maps_only = TRUE, use_phylopics = FALSE,
  point_color = "red", gif_name = NULL)
}
\arguments{
\item{start_time}{The time of the first frame of the animation}

\item{stop_time}{The time of the last frame of the animation before it starts looping back}

\item{periods}{A vector of period names, capitalized properly (can be left blank)}

\item{taxa}{A vector of taxon names (can be left blank)}

\item{step_size}{How many million years to take in a single step}

\item{age_df}{Data.frame of ages, typically from GetAgeDF()}

\item{specimen_df}{Cached fossil localities and times}

\item{interval}{How many seconds per frame}

\item{use_cached_maps_only}{If TRUE, only uses the maps already in the package, rather than pulling from gplates}

\item{use_phylopics}{If TRUE, use phylopic images; otherwise, use dots}

\item{point_color}{If just plotting points, what color}

\item{gif_name}{Path to gif, including its name}
}
\description{
Create an array of animated maps, where the rows are taxa and the columns periods. "all" and "none" are possible taxa, and "all" is a possible period
}
