% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{CacheEverything}
\alias{CacheEverything}
\title{Cache everything}
\usage{
CacheEverything(taxa = GetTaxa(), age_df = GetAgeDF())
}
\arguments{
\item{taxa}{vector of taxa}

\item{age_df}{output of GetAgeDF()}
}
\description{
Just to save typing, run all the caching functions
For all but the images, this won't overwrite if the new object is empty
}
