% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{CacheIndividualTrees}
\alias{CacheIndividualTrees}
\title{Cache trees for website}
\usage{
CacheIndividualTrees(taxa = GetTaxa(), age_df = GetAgeDF(),
  height = 800, width = 900)
}
\arguments{
\item{taxa}{Vector of taxa}

\item{age_df}{From GetAgeDF}

\item{height}{Height in pixels}

\item{width}{Width in pixels}
}
\description{
Cache trees for website
}
