% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{CacheSpecimenAges}
\alias{CacheSpecimenAges}
\title{Cache specimen information}
\usage{
CacheSpecimenAges(taxa = GetTaxa())
}
\arguments{
\item{taxa}{vector of taxa}
}
\description{
Localities of fossils from PBDB
}
