% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{CacheTaxonImages}
\alias{CacheTaxonImages}
\title{Cache taxon information}
\usage{
CacheTaxonImages(taxa = GetTaxa())
}
\arguments{
\item{taxa}{vector of taxa}
}
\description{
Mainly to store images. If an image isn't found for a taxon, an image of DNA is used instead
}
