% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{CacheTaxonImagesByName}
\alias{CacheTaxonImagesByName}
\title{Cache taxon information}
\usage{
CacheTaxonImagesByName(taxa = GetTaxa())
}
\arguments{
\item{taxa}{Vector of taxon names}
}
\description{
Mainly to store images. If an image isn't found for a taxon, an image of DNA is used instead
}
