% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{CacheTaxonImagesFromPhylopic}
\alias{CacheTaxonImagesFromPhylopic}
\title{Cache taxon images from phylopic}
\usage{
CacheTaxonImagesFromPhylopic(tree = chronogram,
  cacheDir = "/Users/bomeara/Documents/MyDocuments/GitClones/HistoryOfEarth/inst/shiny-examples/mainapp/www")
}
\arguments{
\item{tree}{Tree from dateTaxonTreePBDB}

\item{cacheDir}{Where to store trees}
}
\description{
Cache taxon images from phylopic
}
