% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chronogram}
\alias{chronogram}
\title{Saved tree}
\format{A phylo object}
\usage{
data(chronogram)
}
\description{
A phylo object containing a chronogram made with the help of Dave Bapst's paleotree and Paleobiology Database
}
\keyword{datasets}
