% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{CreateMapList}
\alias{CreateMapList}
\title{Make a list of maps for all periods}
\usage{
CreateMapList(age_df = GetAgeDF())
}
\arguments{
\item{age_df}{Output of GetAgeDF()}
}
\value{
list of maps, with names for periods
}
\description{
Make a list of maps for all periods
}
