% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{CreateMapListAllTimes}
\alias{CreateMapListAllTimes}
\title{Make a list of maps for all times}
\usage{
CreateMapListAllTimes(start_age = 0, stop_age = 540, step_size = 1,
  age_df = GetAgeDF(), base_url = "http://gws.gplates.org/")
}
\arguments{
\item{start_age}{How many MYA to start making the map}

\item{stop_age}{How many MYA to stop making the map}

\item{step_size}{How many MY between maps}

\item{age_df}{Output of GetAgeDF()}

\item{base_url}{What URL to use for gplates}
}
\value{
list of maps, with names for periods
}
\description{
By default, uses, gws.gplates.org
}
\details{
However, you can do gplatesr::launch_docker() if you are on a machine with docker running, then use base_url="http://localhost:8888/" to use an instance running locally
}
