% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{get_genuslink}
\alias{get_genuslink}
\title{get link for each genus}
\usage{
get_genuslink(taxon)
}
\arguments{
\item{taxon}{which genus to return a url for}
}
\value{
URL
}
\description{
get link for each genus
}
