% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{get_periodlink}
\alias{get_periodlink}
\title{get link for each period}
\usage{
get_periodlink(period)
}
\arguments{
\item{period}{which period to return a url for}
}
\value{
URL
}
\description{
get link for each period
}
