% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{GetAgeDF}
\alias{GetAgeDF}
\title{Get boundaries and names of geological periods}
\usage{
GetAgeDF()
}
\value{
a data.frame of Periods, min, max, and mid points
}
\description{
Get boundaries and names of geological periods
}
