% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{GetLatLongAnytime}
\alias{GetLatLongAnytime}
\title{Get information on specimens from pbdb}
\usage{
GetLatLongAnytime(taxon)
}
\arguments{
\item{taxon}{a string with a taxon name}
}
\value{
a data.frame with information on this taxon
}
\description{
Get information on specimens from pbdb
}
