% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{GetPreselectedThumbnail}
\alias{GetPreselectedThumbnail}
\title{Get a preselected image}
\usage{
GetPreselectedThumbnail(taxon)
}
\arguments{
\item{taxon}{Taxon to get}
}
\value{
URL of the image
}
\description{
Get a preselected image
}
