% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{GetTree}
\alias{GetTree}
\title{Get tree}
\usage{
GetTree(taxa = GetTaxa(), rank = "genus")
}
\arguments{
\item{taxa}{Vector of taxon names}

\item{rank}{What taxonomic rank to use}
}
\value{
a phylo object
}
\description{
Get tree
}
