% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{GetWikipediaThumbnail}
\alias{GetWikipediaThumbnail}
\title{Get an image from Wikipedia}
\usage{
GetWikipediaThumbnail(taxon)
}
\arguments{
\item{taxon}{Taxon to get}

\item{size}{Size of larger dimension}
}
\value{
URL of the image
}
\description{
Get an image from Wikipedia
}
