% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{latlong_age}
\alias{latlong_age}
\title{Get information on specimens from pbdb}
\usage{
latlong_age(taxa = GetTaxa(), age_df = GetAgeDF())
}
\arguments{
\item{taxa}{a vector of taxon names}
}
\value{
a data.frame with information on all these taxa
}
\description{
Get information on specimens from pbdb
}
