% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paleomaps}
\alias{paleomaps}
\title{Paleomaps}
\format{A list of map objects}
\usage{
data(paleomaps)
}
\description{
Cached maps using LunaSare/gplatesr and gplates, based on example code from https://github.com/GPlates/gplates_web_service_doc by Michael Chin. It uses the https://www.gplates.org/ for its source.
}
\keyword{datasets}
