% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paleomaps_allages}
\alias{paleomaps_allages}
\title{Paleomaps All Ages}
\format{A list of map objects}
\usage{
data(paleomaps_allages)
}
\description{
Cached maps using LunaSare/gplatesr and gplates, based on example code from https://github.com/GPlates/gplates_web_service_doc by Michael Chin. It uses the https://www.gplates.org/ for its source.
}
\details{
Returns one map per 1 MY interval
}
\keyword{datasets}
