% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{PutPointsOnMap}
\alias{PutPointsOnMap}
\title{Put points on a map}
\usage{
PutPointsOnMap(taxa = GetTaxa(), specimen_df = specimens,
  age_df = GetAgeDF(), maps = paleomaps)
}
\arguments{
\item{taxa}{vector of taxon names, default is GetTaxa()}

\item{specimen_df}{data.frame of specimen info}

\item{age_df}{dataframe with Period, MinMa, MaxMa, MidMa, default is GetAgeDF()}

\item{maps}{paleomaps list of maps (stored internally)}
}
\value{
a list of period maps with points for fossils from that period
}
\description{
Put points on a map
}
