% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{recolor_phylopic_for_map}
\alias{recolor_phylopic_for_map}
\title{Fixes error in adding phylopics to maps}
\usage{
recolor_phylopic_for_map(img, alpha = 0.2, color = NULL)
}
\description{
Just copying recolor_phylopic_for_map for now from the rphylopic package
}
