% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{runSite}
\alias{runSite}
\title{Run the interactive website}
\usage{
runSite()
}
\description{
Uses example from https://deanattali.com/2015/04/21/r-package-shiny-app/
}
\details{
In R, just do HistoryOfEarth::runSite()
}
