% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_links}
\alias{taxa_links}
\title{Taxa Links}
\format{A data.frame with info on taxa}
\usage{
data(taxa_links)
}
\description{
Selected taxa common names, scientific names, links to more info, links to picture
}
\keyword{datasets}
