% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonimages}
\alias{taxonimages}
\title{Taxon Images}
\format{A list of images with names equal to the taxon names}
\usage{
data(taxonimages)
}
\description{
Images of taxa from phylopic
}
\keyword{datasets}
