% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{CacheMapsAllAges}
\alias{CacheMapsAllAges}
\title{Cache map information all ages}
\usage{
CacheMapsAllAges(base_url = "http://gws.gplates.org/")
}
\arguments{
\item{base_url}{What URL to use for gplates}
}
\description{
Cache map information all ages
}
