% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{CacheTaxonImagesFromPhylopic}
\alias{CacheTaxonImagesFromPhylopic}
\title{Cache taxon images from phylopic}
\usage{
CacheTaxonImagesFromPhylopic(tree = chronogram, cacheDir = "docs/img")
}
\arguments{
\item{tree}{Tree from dateTaxonTreePBDB}

\item{cacheDir}{Where to store trees}
}
\description{
Cache taxon images from phylopic
}
