% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{get_pictree}
\alias{get_pictree}
\title{Get phylopic tree}
\usage{
get_pictree(tree = GetTree())
}
\arguments{
\item{taxa}{Vector of taxon names, default is GetTree()}
}
\value{
a plotted tree with phylopics
}
\description{
Get phylopic tree
}
