% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{get_pictree}
\alias{get_pictree}
\title{Get phylopic tree}
\usage{
get_pictree(tree = NULL, cacheDir = "./www/")
}
\arguments{
\item{tree}{phylo object}

\item{cacheDir}{where to load images from if they exist}
}
\value{
a plotted tree with phylopics
}
\description{
If no tree is provided, will use the cached tree
}
