% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historyofearth.R
\name{GetGBIFPoints}
\alias{GetGBIFPoints}
\title{Get locations for modern species using GBIF}
\usage{
GetGBIFPoints(taxa = GetTaxa(), rank = "genus")
}
\arguments{
\item{taxa}{vector of taxon names, default is GetTaxa()}

\item{the}{rank to search at for each taxon}
}
\value{
data frame with pbdb_data.paleolng and pbdb_data.paleolat columns reflecting points from gbif
}
\description{
Note that to make this work with the add_points function, needs odd output column names
}
